remove BaseClient class in favor of a single Client class
